Restructured requestUserConfirm modal to return boolean for confirmed yes/no. In most cases we can use this boolean instead of relying on errors.

+++ src/Eagle.ts
@@ -2304,13 +2304,16 @@ export class Eagle {
 
                 // warn user if file newer than EAGLE
                 if (Utils.newerEagleVersion(eagleVersion, (<any>window).version)){
+                    let confirmed: boolean = false;
                     try {
-                        await Utils.requestUserConfirm("Newer EAGLE Version", "File " + file.name + " was written with EAGLE version " + eagleVersion + ", whereas the current EAGLE version is " + (<any>window).version + ". Do you wish to load the file anyway?", "Yes", "No", null);
+                        confirmed = await Utils.requestUserConfirm("Newer EAGLE Version", "File " + file.name + " was written with EAGLE version " + eagleVersion + ", whereas the current EAGLE version is " + (<any>window).version + ". Do you wish to load the file anyway?", "Yes", "No", null);
                     } catch (error){
                         console.error(error);
                         return;
                     }
-                    this._loadGraph(dataObject, file);
+                    if (confirmed){
+                        this._loadGraph(dataObject, file);
+                    }
                 } else {
                     this._loadGraph(dataObject, file);
                 }
@@ -2450,13 +2453,16 @@ export class Eagle {
 
     deleteRemoteFile = async (file : RepositoryFile): Promise<void> => {
         // request confirmation from user
+        let confirmed: boolean = false;
         try {
-            await Utils.requestUserConfirm("Delete?", "Are you sure you wish to delete '" + file.name + "' from this repository?", "Yes", "No", Setting.find(Setting.CONFIRM_DELETE_FILES));
+            confirmed = await Utils.requestUserConfirm("Delete?", "Are you sure you wish to delete '" + file.name + "' from this repository?", "Yes", "No", Setting.find(Setting.CONFIRM_DELETE_FILES));
         } catch (error) {
             console.error(error);
             return;
         }
-        this._deleteRemoteFile(file);
+        if (confirmed) {
+            this._deleteRemoteFile(file);
+        }
     }
 
     private _deleteRemoteFile = async (file: RepositoryFile): Promise<void> => {
@@ -2499,13 +2505,16 @@ export class Eagle {
 
         // if dictated by settings, reload the palette immediately
         if (alreadyLoadedPalette !== null && Setting.findValue(Setting.CONFIRM_RELOAD_PALETTES)){
+            let confirmed: boolean = false;
             try {
-                await Utils.requestUserConfirm("Reload Palette?", "This palette (" + file.name + ") is already loaded, do you wish to load it again?", "Yes", "No", Setting.find(Setting.CONFIRM_RELOAD_PALETTES));
+                confirmed = await Utils.requestUserConfirm("Reload Palette?", "This palette (" + file.name + ") is already loaded, do you wish to load it again?", "Yes", "No", Setting.find(Setting.CONFIRM_RELOAD_PALETTES));
             } catch (error){
                 console.error(error);
                 return;
             }
-            this._reloadPalette(file, data, alreadyLoadedPalette);
+            if (confirmed){
+                this._reloadPalette(file, data, alreadyLoadedPalette);
+            }
         } else {
             this._reloadPalette(file, data, alreadyLoadedPalette);
         }
@@ -2575,13 +2584,16 @@ export class Eagle {
 
                 // check if the palette is modified, and if so, ask the user to confirm they wish to close
                 if (p.fileInfo().modified && Setting.findValue(Setting.CONFIRM_DISCARD_CHANGES)){
+                    let confirmed: boolean = false;
                     try {
-                        await Utils.requestUserConfirm("Close Modified Palette", "Are you sure you wish to close this modified palette?", "Close", "Cancel", null);
+                        confirmed = await Utils.requestUserConfirm("Close Modified Palette", "Are you sure you wish to close this modified palette?", "Close", "Cancel", null);
                     } catch (error){
                         console.error(error);
                         return;
                     }
-                    this.palettes.splice(i, 1);
+                    if (confirmed){
+                        this.palettes.splice(i, 1);
+                    }
                 } else {
                     this.palettes.splice(i, 1);
                 }
@@ -3531,14 +3543,17 @@ export class Eagle {
         }
 
         // request confirmation from user
+        let confirmed: boolean = false;
         try {
-            await Utils.requestUserConfirm("Delete?", confirmMessage, "Yes", "No", Setting.find(Setting.CONFIRM_DELETE_OBJECTS));
+            confirmed = await Utils.requestUserConfirm("Delete?", confirmMessage, "Yes", "No", Setting.find(Setting.CONFIRM_DELETE_OBJECTS));
         } catch (error) {
             console.error(error);
             return;
         }
 
-        this._deleteSelection(deleteChildren, data, location);
+        if (confirmed){
+            this._deleteSelection(deleteChildren, data, location);
+        }
 
         // if we're NOT in rightClick mode, empty the selected objects, should have all been deleted
         if(!rightClick){
@@ -4570,14 +4585,17 @@ export class Eagle {
         // request confirmation from user
         // old request if 'confirm' setting is true AND we're not going to keep the old fields
         if (confirmNodeCategoryChanges && !keepOldFields){
+            let confirmed: boolean = false;
             try {
-                await Utils.requestUserConfirm("Change Category?", 'Changing a nodes category could destroy some data (parameters, ports, etc) that are not appropriate for a node with the selected category', "Yes", "No", Setting.find(Setting.CONFIRM_NODE_CATEGORY_CHANGES));
+                confirmed = await Utils.requestUserConfirm("Change Category?", 'Changing a nodes category could destroy some data (parameters, ports, etc) that are not appropriate for a node with the selected category', "Yes", "No", Setting.find(Setting.CONFIRM_NODE_CATEGORY_CHANGES));
             } catch (error){
                 //we need to reset the input select to the previous value
                 $(event.target).val(this.selectedNode().getCategory())
                 return;
             }
-            this.inspectorChangeNodeCategory(event)
+            if (confirmed){
+                this.inspectorChangeNodeCategory(event)
+            }
         }else{
             this.inspectorChangeNodeCategory(event)
         }
